Only keep connection while reading/writing
Before the connections was opened as objects were instantiated.
This made it so concurrent writes would timeout for long running jobs.
Now the connection is opened and closed on every insertion, deletion, or query.
The purpose is not to have high performance operations, but to allow for
many concurrent workers to experiment results in a common database.